+++ src/tooling/piral-cli-webpack5/src/enhancers/pilet-webpack-config-enhancer/helpers.ts
@@ -77,7 +77,7 @@ export function withSetPath(compilerOptions: Configuration) {
 }
 
 export function withExternals(compilerOptions: Configuration, externals: Array<string>) {
-  const current = compilerOptions.externals;
+  const current = compilerOptions.externals || [];
   const arrayExternals = Array.isArray(current) ? current : [current];
 
   const objectExternal = externals.reduce((external, dep) => {