+++ src/tooling/piral-cli/src/common/url.ts
@@ -1,0 +1,15 @@
+export function normalizePublicUrl(url: string) {
+  if (!url) {
+    return '/';
+  }
+
+  if (!url.endsWith('/')) {
+    url = `${url}/`;
+  }
+
+  if (!url.startsWith('/')) {
+    url = `/${url}`;
+  }
+
+  return url;
+}