+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Piral Changelog
 
+## 0.14.9 (tbd)
+
+- Improved error output in `piral-cli-webpack5`
+- Improved error handling in `parcel-codegen-loader`
+- Improved `publicUrl` argument with normalization
+- Fixed types section missing in documentation
+- Fixed error with externals in `piral-cli-webpack`
+- Fixed error with externals in `piral-cli-webpack5`
+
 ## 0.14.8 (January 20, 2022)
 
 - Improved performance of internal action calls